Fix für Offline Aufgaben 2 3

+++ Offline Aufgaben/2/3/main.cpp
@@ -17,8 +17,8 @@ int main() {
 	kandidat3 = summe - kandidat1 - kandidat2;
 	cout << "Auf den dritten Kandidaten sind somit " << kandidat3 << " Stimmen entfallen." << endl;
 
-	cout << "Kandidat 1 erhielt " << float(kandidat1) / summe * 100 << "\% der Stimmen." << endl;
-	cout << "Kandidat 2 erhielt " << float(kandidat2) / summe * 100 << "\% der Stimmen." << endl;
-	cout << "Kandidat 3 erhielt " << float(kandidat3) / summe * 100 << "\% der Stimmen." << endl;
+	cout << "Kandidat 1 erhielt " << double(kandidat1) / summe * 100 << "\% der Stimmen." << endl;
+	cout << "Kandidat 2 erhielt " << double(kandidat2) / summe * 100 << "\% der Stimmen." << endl;
+	cout << "Kandidat 3 erhielt " << double(kandidat3) / summe * 100 << "\% der Stimmen." << endl;
 	return 0;
 }
